1,clear warnings 2,add code license to the pppos 3,move pppos to component
