+++ components/connectivity/agent_tiny/examples/agent_tiny_demo.c
@@ -33,27 +33,36 @@
  *---------------------------------------------------------------------------*/
 
 #include "agent_tiny_demo.h"
-#ifdef CONFIG_FEATURE_FOTA
 #include "fota_port.h"
-#endif
 
-#define DEFAULT_SERVER_IPV4 "180.101.147.115" /*dianxin*/
+#ifdef CONFIG_FEATURE_FOTA
+#endif
+//#define DEFAULT_SERVER_IPV4 "139.159.209.89"/*Huawei */
+//#define DEFAULT_SERVER_IPV4 "192.168.0.116"/*Huawei */
+//#define DEFAULT_SERVER_IPV4 "192.168.1.106"/*sjn */
+#define DEFAULT_SERVER_IPV4 "139.159.140.34" /*czr*/
 
 #define LWM2M_LIFE_TIME     50000
 
-#define IOT_PSK_VALUE_LENGTH    16
-#define BS_PSK_VALUE_LENGTH     16
-char * g_endpoint_name = "44440003";
+#define IOT_PSK_VALUE_LENGTH    12
+#define BS_PSK_VALUE_LENGTH     12
+char * g_endpoint_name = "power08";
 #ifdef WITH_DTLS
+//char *g_endpoint_name_s = "power07";
+//unsigned char g_psk_value[16]={ 0xb7, 0xed, 0x6c, 0xd8, 0x6b, 0x78, 0x29, 0x3f, 0x35, 0x11, 0x6d, 0x8b, 0x18, 0x87, 0x6b, 0x10};
 
-char* g_endpoint_name_s = "55555002";
-char* g_endpoint_name_iots = "55555002";
-char* g_endpoint_name_bs = "55555002";
-unsigned char g_psk_iot_value[IOT_PSK_VALUE_LENGTH] = {0xb4,0x72,0x8f,0x8d,0x70,0xbd,0xfc,0x35,0xf1,0xed,0xf5,0x2b,0xf6,0xa1,0x98,0xf7};  //0x33 -> 0x32
-unsigned char g_psk_bs_value[BS_PSK_VALUE_LENGTH] = {0xb4,0x72,0x8f,0x8d,0x70,0xbd,0xfc,0x35,0xf1,0xed,0xf5,0x2b,0xf6,0xa1,0x98,0xf7};
-//unsigned char g_psk_value[16] = {0xef,0xe8,0x18,0x45,0xa3,0x53,0xc1,0x3c,0x0c,0x89,0x92,0xb3,0x1d,0x6b,0x6a,0x33};
+//char *g_endpoint_name_s = "88889999";
+//unsigned char g_psk_value[16] = {0x02,0x77,0x68,0xca,0x0b,0xf5,0xdf,0xba,0x46,0x43,0x25,0xdd,0x4b,0xe7,0x0a,0x9d};
 
+char* g_endpoint_name_s = "power07";
+char* g_endpoint_name_iots = "power07";
+char* g_endpoint_name_bs = "power07";
+unsigned char g_psk_iot_value[IOT_PSK_VALUE_LENGTH] = { 0x6b, 0x78, 0x29, 0x3f, 0x35, 0x11, 0x6d, 0x8b, 0x18, 0x87, 0x6b, 0x10};  //0x33 -> 0x32
+unsigned char g_psk_bs_value[BS_PSK_VALUE_LENGTH] = { 0x6b, 0x78, 0x29, 0x3f, 0x35, 0x11, 0x6d, 0x8b, 0x18, 0x87, 0x6b, 0x10};
+unsigned char g_psk_value[16]={ 0xb7, 0xed, 0x6c, 0xd8, 0x6b, 0x78, 0x29, 0x3f, 0x35, 0x11, 0x6d, 0x8b, 0x18, 0x87, 0x6b, 0x10};
 #endif
+
+#define CN_DTLS_SERVER   1
 
 static void* g_phandle = NULL;
 static atiny_device_info_t g_device_info;
@@ -61,33 +70,33 @@ static atiny_param_t g_atiny_params;
 
 void ack_callback(atiny_report_type_e type, int cookie, data_send_status_e status)
 {
-    ATINY_LOG(LOG_DEBUG,"type:%d cookie:%d status:%d\n", type,cookie, status);
+    printf("type:%d cookie:%d status:%d\n", type,cookie, status);
 }
-/*lint -e550*/
+
+
+unsigned int checkstack = 0;
 void app_data_report(void)
 {
     uint8_t buf[5] = {0,1,6,5,9};
     data_report_t report_data;
-    int ret = 0;
+    int ret;
     int cnt = 0;
     report_data.buf = buf;
     report_data.callback = ack_callback;
     report_data.cookie = 0;
     report_data.len = sizeof(buf);
     report_data.type = APP_DATA;
-    (void)ret;
     while(1)
     {
         report_data.cookie = cnt;
         cnt++;
         ret = atiny_data_report(g_phandle, &report_data);
-        ATINY_LOG(LOG_DEBUG,"data report ret: %d\n",ret);
+        printf("report ret: %d\n",ret);
         ret = atiny_data_change(g_phandle, DEVICE_MEMORY_FREE);
-        ATINY_LOG(LOG_DEBUG,"data change ret: %d\n",ret);
-        (void)LOS_TaskDelay(250*8);
+        printf("data change ret: %d\n",ret);
+        (void)LOS_TaskDelay(1000*10);
     }
 }
-/*lint +e550*/
 
 UINT32 creat_report_task()
 {
@@ -98,7 +107,7 @@ UINT32 creat_report_task()
     task_init_param.usTaskPrio = 1;
     task_init_param.pcName = "app_data_report";
     task_init_param.pfnTaskEntry = (TSK_ENTRY_FUNC)app_data_report;
-    task_init_param.uwStackSize = 0x1000;
+    task_init_param.uwStackSize = 768;
 
     uwRet = LOS_TaskCreate(&TskHandle, &task_init_param);
     if(LOS_OK != uwRet)
@@ -133,7 +142,7 @@ void agent_tiny_entry(void)
     atiny_param_t* atiny_params;
     atiny_security_param_t  *iot_security_param = NULL;
     atiny_security_param_t  *bs_security_param = NULL;
-
+    atiny_security_param_t  *security_param = NULL;
     atiny_device_info_t *device_info = &g_device_info;
 
 #ifdef CONFIG_FEATURE_FOTA
@@ -145,12 +154,8 @@ void agent_tiny_entry(void)
 #else
     device_info->endpoint_name = g_endpoint_name;
 #endif
-#ifdef CONFIG_FEATURE_FOTA
-    device_info->manufacturer = "Lwm2mFota";
-    device_info->dev_type = "Lwm2mFota";
-#else
     device_info->manufacturer = "Agent_Tiny";
-#endif
+
     atiny_params = &g_atiny_params;
     atiny_params->server_params.binding = "UQ";
     //atiny_params->server_params.life_time = LWM2M_LIFE_TIME;
@@ -168,6 +173,15 @@ void agent_tiny_entry(void)
     bs_security_param->server_ip = DEFAULT_SERVER_IPV4;
 
 #ifdef WITH_DTLS
+
+#ifdef CN_DTLS_SERVER
+    security_param = &(atiny_params->security_params[0]);
+    security_param->server_ip = DEFAULT_SERVER_IPV4;
+    security_param->server_port = "5684";
+    security_param->psk_Id = g_endpoint_name_s;
+    security_param->psk = (char*)g_psk_value;
+    security_param->psk_len = 16;
+#else
     iot_security_param->server_port = "5684";
     bs_security_param->server_port = "5684";
 
@@ -178,6 +192,7 @@ void agent_tiny_entry(void)
     bs_security_param->psk_Id = g_endpoint_name_bs;
     bs_security_param->psk = (char*)g_psk_bs_value;
     bs_security_param->psk_len = BS_PSK_VALUE_LENGTH;
+#endif
 #else
     iot_security_param->server_port = "5683";
     bs_security_param->server_port = "5683";
